Rename doc files
