+++ doc/pricing.rb
