+++ doc/ruby_array.rb
